% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficientvariation.R
\name{CoefVar}
\alias{CoefVar}
\title{Estimate standard error or coefficient of variation}
\usage{
CoefVar(x)
}
\arguments{
\item{x}{vector of data}
}
\description{
Returns standard error around vector of data x.
}
\examples{
mydata <- c(1, 2, 1, 6, 7, 8, 10)
CoefVar(mydata)
}

