\docType{package}
\name{DESutils}
\alias{DESutils}
\alias{DESutils-package}
\title{Functions for discrete event simulation in health economics}
\description{
  Provides tools that make it easier to perform and and
  analyse the output of discrete event simulation.
  Specifically in the context of health economic
  evaluation. More information about each function can be
  found in its help documentation.
}
\author{
  David Trueman <david_trueman@hotmail.co.uk>
}
\references{
  Law, A. Simulation Modeling and Analysis, Fourth Edition.
  McGraw-Hill, 2007
  \url{http://www.averill-law.com/simulation-book/}
}

