% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guyot.R
\name{DigitizeKM}
\alias{DigitizeKM}
\title{Create patient-level dataset from published KM curves}
\usage{
DigitizeKM(my.path, n.arms = 1, surv.data, risk.data, tot.events = NULL,
  out.path = NULL)
}
\arguments{
\item{my.path}{string of file from where all the data must be accessible}

\item{n.arms}{integer representing teh number of arms in the study/data}

\item{surv.data}{vector of strings containing names of csv data files with
times and probability surviving. One per arm must be supplied. Note:
filepath and the strings in this vector will be combined to retrieve data.}

\item{risk.data}{vector of strings containing names of csv data files with
times intervals, corresponding points from surv.data, and n at risk. One per arm must be supplied. Note:
filepath and the strings in this vector will be combined to retrieve data.}

\item{tot.events}{total number of events reported. If not reported, then tot.events=NULL.
If included, should be a list with number of elements = n.arms}

\item{out.path}{a string representing a filepath where temporaryd datasets will be stored.
If not specified, filepath will be used}
}
\description{
Wrapper for algorithm provided by Guyolt et al. to facilitate easier use.
Resturns a patient-level dataset
}
\examples{
#Original 'real' data
require(survival)
bladder
asurv<-Surv(bladder$stop, bladder$event)
afit<-survfit(asurv~bladder$rx)
plot(afit)

#Reconstructed data
patient.data <- DigitizeKM(my.path="data/bladder",
                          n.arms=2,
                          surv.data=c('Treatment1.csv', 'Treatment2.csv'),
                          risk.data=c('tx1_risk.csv','tx2_risk.csv'))
                          
                          
bfit<-survfit(Surv(time, fail)~arm, data=patient.data)
plot(bfit)
}
\references{
Guyot et al. Enhanced secondary analysis of survival data: reconstructing the data from published Kaplan-Meier survival curves. BMC Medical Research Methodology 2012, 12:9
}

