% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findit.R
\name{FindIt}
\alias{FindIt}
\title{Return Index of value from vector}
\usage{
FindIt(x, vec)
}
\arguments{
\item{x}{value to look up}

\item{vec}{vector of data to look up from}
}
\description{
Returns a row number corresponding to the closest value within a vector of data
}

