\name{InverseGamma}
\alias{InverseGamma}
\title{Returns inverse gamma distribution}
\usage{
  InverseGamma(alpha, beta = 1)
}
\arguments{
  \item{alpha}{alpha coefficient}

  \item{beta}{beta coefficient}
}
\description{
  Generate inverse gamma distribution based on algorithms
  on p. 451 of Law, A. Simulation modelling and analysis.
  4th Ed. 2007
}

