\name{PresentValueOne}
\alias{PresentValueOne}
\title{Return present value of a one-off cost in the future}
\usage{
  PresentValueOne(value.to.discount, discount.rate,
    time.point = 0)
}
\arguments{
  \item{value.to.discount}{value of one-off event}

  \item{discount.rate}{discount rate e.g. 0.035}

  \item{time.point}{time at which event occurs (in years)}
}
\description{
  Reurns present value of a one-off cost at some point in
  the future
}

