% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refcostsumweights.R
\name{SumWeights}
\alias{SumWeights}
\title{Returns weighted lower or upper quartile for selection of NHS ref costs}
\usage{
SumWeights(cut, data, low = TRUE)
}
\arguments{
\item{cut}{cut-point on log-scale}

\item{data}{dataset in .csv format to use with 9 columns of data, 1 row per currency code}

\item{low}{low=TRUE if we wish to find the cut-point corresponding to 25th percentile,
or low=FALSE if we wish to fnd the cut-point corresponding to 75th percentile}
}
\description{
Normally called by RefCostsRange. In order to estimate the appropriate weighting
of lower/upper quartiles a 'cut-point' is used.
}

